requirements.txt
asgiref==3.7.2
attrs==23.2.0
Babel==2.10.3
blinker==1.7.0
celery==5.4.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloudinary==1.44.1
colorama==0.4.6
cryptography==41.0.7
dj-database-url==3.0.1
Django==4.2.11
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
h11==0.14.0
idna==3.6
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
netaddr==0.8.0
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.10
Pygments==2.17.2
PyJWT==2.7.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.1.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
rich==13.7.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.0.7
uvicorn==0.27.1
uvloop==0.19.0
wadllib==1.3.6
wheel==0.42.0
wsproto==1.2.0
gspread==6.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1